requirements.txt
numpy
pandas
streamlit
scikit-learn
python-dateutil
nltk